Revert "feat: Private Figma files"
This reverts commit fb311147195b1f8d13dbbc33b1c77b2ee1d5f5d6.
