--- packages/storybook-addon-designs/oauth.html
@@ -1,22 +1,0 @@
-<html>
-  <head></head>
-  <body>
-    <script>
-      const url = new URL(location.href)
-
-      window.opener.postMessage(
-        JSON.stringify({
-          type: 'OAUTH_COMPLETE',
-          payload: {
-            code: url.searchParams.get('code'),
-            state: url.searchParams.get('state')
-          }
-        }),
-        location.origin
-      )
-
-      console.log(url.searchParams.get('code'))
-      console.log(window.opener)
-    </script>
-  </body>
-</html>
