+++ packages/storybook-addon-designs/package.json
@@ -21,8 +21,6 @@
     "@storybook/components": "^5.0.3",
     "@storybook/core-events": "^5.0.3",
     "@storybook/theming": "^5.0.3",
-    "@types/js-cookie": "^2.2.5",
-    "@types/object-hash": "^1.3.1",
     "@types/pdfjs-dist": "^2.0.0",
     "@types/react": "^16.8.8",
     "@types/webpack-env": "^1.13.9",
@@ -45,9 +43,6 @@
     "prepublishOnly": "cp ../../README.md ./ && yarn build"
   },
   "dependencies": {
-    "html-webpack-plugin": "^3.2.0",
-    "js-cookie": "^2.2.1",
-    "object-hash": "^2.0.3",
     "react-pdf": "^4.0.5"
   },
   "publishConfig": {
