+++ packages/storybook-addon-designs/preset.js
@@ -1,7 +1,3 @@
-const path = require('path')
-
-const HtmlPlugin = require('html-webpack-plugin')
-
 module.exports = {
   managerEntries(entries = [], options = {}) {
     return [
@@ -10,18 +6,5 @@ module.exports = {
         ? require.resolve('./register-tab')
         : require.resolve('./register')
     ]
-  },
-  async webpack(config) {
-    return {
-      ...config,
-      plugins: [
-        ...(config.plugins || []),
-        new HtmlPlugin({
-          inject: false,
-          filename: 'addon-designs-oauth.html',
-          template: path.resolve(__dirname, './oauth.html')
-        })
-      ]
-    }
   }
 }
